+++ tailwind.config.ts
@@ -12,10 +12,10 @@ module.exports = {
     extend: {
       colors: {
         primary: {
-          50: '#fef2f2',
-          500: '#ef4444',
-          600: '#dc2626',
-          700: '#b91c1c',
+          // 50: '#fef2f2',
+          // 500: '#ef4444',
+          // 600: '#dc2626',
+          // 700: '#b91c1c',
         }
       }
     },
